refactor: :recycle: move renderLore and formatNumber into common
