+++ common/constants.js
@@ -1,0 +1,126 @@
+export const specialEnchants = [
+  "Sharpness 7",
+  "Sharpness VII",
+  "Sharpness 6",
+  "Sharpness VI",
+  "Giant Killer 7",
+  "Giant Killer VII",
+  "Giant Killer 6",
+  "Giant Killer VI",
+  "Bane of Arthropods 7",
+  "Bane of Arthropods VII",
+  "Bane of Arthropods 6",
+  "Bane of Arthropods VI",
+  "Critical 7",
+  "Critical VII",
+  "Critical 6",
+  "Critical VI",
+  "Ender Slayer 7",
+  "Ender Slayer VII",
+  "Ender Slayer 6",
+  "Ender Slayer VI",
+  "Experience 4",
+  "Experience IV",
+  "Life Steal 5",
+  "Life Steal V",
+  "Life Steal 4",
+  "Life Steal IV",
+  "Looting 5",
+  "Looting V",
+  "Looting 4",
+  "Looting IV",
+  "Luck 7",
+  "Luck VII",
+  "Luck 6",
+  "Luck VI",
+  "Scavenger 5",
+  "Scavenger V",
+  "Scavenger 4",
+  "Scavenger IV",
+  "Smite 7",
+  "Smite VII",
+  "Smite 6",
+  "Smite VI",
+  "Vampirism 6",
+  "Vampirism VI",
+  "Power 7",
+  "Power VII",
+  "Power 6",
+  "Power VI",
+  "Growth 7",
+  "Growth VII",
+  "Growth 6",
+  "Growth VI",
+  "Blast Protection 7",
+  "Blast Protection VII",
+  "Blast Protection 6",
+  "Blast Protection VI",
+  "Fire Protection 7",
+  "Fire Protection VII",
+  "Fire Protection 6",
+  "Fire Protection VI",
+  "Projectile Protection 7",
+  "Projectile Protection VII",
+  "Projectile Protection 6",
+  "Projectile Protection VI",
+  "Protection 7",
+  "Protection VII",
+  "Protection 6",
+  "Protection VI",
+  "True Protection 1",
+  "True Protection I",
+  "Sugar Rush 3",
+  "Sugar Rush III",
+  "Efficiency 6",
+  "Efficiency VI",
+  "Angler 6",
+  "Angler VI",
+  "Caster 6",
+  "Caster VI",
+  "Frail 6",
+  "Frail VI",
+  "Luck of the Sea 6",
+  "Luck of the Sea VI",
+  "Lure 6",
+  "Lure VI",
+  "Magnet 6",
+  "Magnet VI",
+  "Spiked Hook 6",
+  "Spiked Hook VI",
+  "Dragon Hunter 5",
+  "Dragon Hunter V",
+  "Feather Falling 10",
+  "Feather Falling X",
+  "Feather Falling 20",
+  "Feather Falling XX",
+  "Infinite Quiver 10",
+  "Infinite Quiver X",
+  "Thorns 6",
+  "Thorns VI",
+  "Expertise 10",
+  "Expertise X",
+  "Thunderlord 6",
+  "Thunderlord VI",
+  "Lethality 6",
+  "Lethality VI",
+  "Compact 10",
+  "Compact X",
+  "Cultivating 10",
+  "Cultivating X",
+  "First Strike 5",
+  "First Strike V",
+  "Execute 6",
+  "Execute VI",
+  "Cubism 6",
+  "Cubism VI",
+  "Venomous 6",
+  "Venomous VI",
+  "Cleave 6",
+  "Cleave VI",
+  "Harvesting 6",
+  "Harvesting VI",
+  "Fortune 4",
+  "Fortune IV",
+  "Pristine 5",
+  "Pristine V",
+];
