+++ src/helper.js
@@ -5,6 +5,8 @@ import "axios-debug-log";
 import { v4 } from "uuid";
 import retry from "async-retry";
 
+export { renderLore, formatNumber } from "../common/formatting.js";
+
 import * as constants from "./constants.js";
 import credentials from "./credentials.js";
 
@@ -355,67 +357,6 @@ export function getGuildLevel(xp) {
 }
 
 /**
- * Convert Minecraft lore to HTML
- * @param {string} text minecraft lore with color and formatting codes
- * @returns {string} HTML
- */
-export function renderLore(text) {
-  let output = "";
-
-  /**
-   * @typedef {"0"|"1"|"2"|"3"|"4"|"5"|"6"|"7"|"8"|"9"|"a"|"b"|"c"|"d"|"e"|"f"} ColorCode
-   * @typedef {"k"|"l"|"m"|"n"|"o"} FormatCode
-   */
-
-  /** @type {ColorCode|null} */
-  let color = null;
-  /** @type {Set<FormatCode>} */
-  let formats = new Set();
-
-  for (let part of text.match(/(§[0-9a-fk-or])*[^§]*/g)) {
-    while (part.charAt(0) === "§") {
-      const code = part.charAt(1);
-
-      if (/[0-9a-f]/.test(code)) {
-        color = code;
-      } else if (/[k-o]/.test(code)) {
-        formats.add(code);
-      } else if (code === "r") {
-        color = null;
-        formats.clear();
-      }
-
-      part = part.substring(2);
-    }
-
-    if (part.length === 0) continue;
-
-    output += "<span";
-
-    if (color !== null) {
-      output += ` style='color: var(--§${color});'`;
-    }
-
-    if (formats.size > 0) {
-      output += ` class='${Array.from(formats, (x) => "§" + x).join(" ")}'`;
-    }
-
-    output += `>${part}</span>`;
-  }
-
-  const matchingEnchants = constants.special_enchants.filter((a) => output.includes(a));
-
-  for (const enchantment of matchingEnchants) {
-    if (enchantment == "Power 6" || (enchantment == "Power 7" && text.startsWith("§8Breaking"))) {
-      continue;
-    }
-    output = output.replace(enchantment, `<span style='color: var(--§6)'>${enchantment}</span>`);
-  }
-
-  return output;
-}
-
-/**
  * Get Minecraft lore without the color and formatting codes
  * @param {string} text lore with color codes
  * @returns {string} lore without color codes
@@ -524,49 +465,6 @@ export function getPrices(product) {
     sellPrice: getPrice(product.sell_summary),
   };
 }
-
-/**
- * @param {number} number the number to be formatted
- * @param {boolean} floor rounds down if true, up if false
- * @param {number} rounding power of ten of the number of digits you want after the decimal point
- *
- * @example formatNumber(123456798, true, 10) = "123.4M"
- * @example formatNumber(123456798, true, 100) = "123.45M"
- */
-export const formatNumber = (number, floor, rounding = 10) => {
-  if (number < 1000) {
-    return Math.floor(number);
-  } else if (number < 10000) {
-    if (floor) {
-      return (Math.floor((number / 1000) * rounding) / rounding).toFixed(rounding.toString().length - 1) + "K";
-    } else {
-      return (Math.ceil((number / 1000) * rounding) / rounding).toFixed(rounding.toString().length - 1) + "K";
-    }
-  } else if (number < 1000000) {
-    if (floor) {
-      return Math.floor(number / 1000) + "K";
-    } else {
-      return Math.ceil(number / 1000) + "K";
-    }
-  } else if (number < 1000000000) {
-    if (floor) {
-      return (Math.floor((number / 1000 / 1000) * rounding) / rounding).toFixed(rounding.toString().length - 1) + "M";
-    } else {
-      return (Math.ceil((number / 1000 / 1000) * rounding) / rounding).toFixed(rounding.toString().length - 1) + "M";
-    }
-  } else if (floor) {
-    return (
-      (Math.floor((number / 1000 / 1000 / 1000) * rounding * 10) / (rounding * 10)).toFixed(
-        rounding.toString().length
-      ) + "B"
-    );
-  } else {
-    return (
-      (Math.ceil((number / 1000 / 1000 / 1000) * rounding * 10) / (rounding * 10)).toFixed(rounding.toString().length) +
-      "B"
-    );
-  }
-};
 
 /**
  * calculates the letter grade of a dungeon Run
