+++ views/stats.ejs
@@ -2719,7 +2719,6 @@ const metaDescription = getMetaDescription()
     const calculated = JSON.parse(`<%- JSON.stringify(calculated).replaceAll('\\', '\\\\') %>`);
     <%
       const clientConstants = {
-        special_enchants: constants.special_enchants,
         max_favorites: constants.max_favorites,
       }
     %>
